When the PDF download option is enabled, delete the PDF after the 2nd time loading
Some browsers will give users the cached copy of the PDF when using the download feature.
Others will try download it from the source instead (including Chrome on iPhone and Windows Safari). This PR allows users who have the
PDF configured for download to access the document from the source one time. It'll be deleted the 2nd time it's accessed.